test(tts): share polling helper between test_support and integration tests
The poll-until idiom existed twice: test_support::wait_until (private
to unit tests) and two hand-rolled loop+sleep+timeout waits in
tests/supervisor.rs. Moved wait_until into tts::supervisor::test_helpers
— the module already gated on cfg(any(test, feature = "test-helpers"))
for exactly this sharing — re-exported it from test_support so command
tests keep a single import site, and replaced the manual loops in the
supervisor integration tests with it.
